feat: add ts enum
