+++ src/components/controller/loader.ts
@@ -1,0 +1,52 @@
+import { IOptions } from '../../types/index';
+import { HttpMethods } from '../../types/enum';
+
+class Loader {
+  private baseLink: string;
+  private options: IOptions;
+
+  constructor(baseLink: string, options: IOptions) {
+    this.baseLink = baseLink;
+    this.options = options;
+  }
+
+  protected getResp<T>(
+    { endpoint = 'everything', options = {} },
+    callback: (data: T) => void = () => {
+      console.error('No callback for GET response');
+    }
+  ) {
+    this.load(HttpMethods.GET, endpoint, callback, options);
+  }
+
+  private errorHandler(res: Response) {
+    if (!res.ok) {
+      if (res.status === 401 || res.status === 404)
+        console.log(`Sorry, but there is ${res.status} error: ${res.statusText}`);
+      throw Error(res.statusText);
+    }
+
+    return res;
+  }
+
+  makeUrl(options: IOptions, endpoint: string): string {
+    const urlOptions: IOptions = { ...this.options, ...options };
+    let url = `${this.baseLink}${endpoint}?`;
+
+    Object.keys(urlOptions).forEach((key) => {
+      url += `${key}=${urlOptions[key]}&`;
+    });
+
+    return url.slice(0, -1);
+  }
+
+  private load<T>(method: HttpMethods, endpoint: string, callback: (data: T) => void, options = {}) {
+    fetch(this.makeUrl(options, endpoint), { method })
+      .then(this.errorHandler)
+      .then((res) => res.json())
+      .then((data) => callback(data))
+      .catch((err) => console.error(err));
+  }
+}
+
+export default Loader;
