+++ src/types/enum.ts
@@ -1,0 +1,4 @@
+export enum HttpMethods {
+  GET = 'GET',
+  POST = 'POST',
+}
